fix: move SequencePackingGradientTestActor to separate module to avoid pytest import in mcore worker venv
